moved Encoder and Decoder into new Codec object
