+++ dynamodb/src/main/scala/zio/dynamodb/AttributeValue.scala
@@ -5,7 +5,7 @@ import zio.dynamodb.ConditionExpression._
 import zio.schema.Schema
 
 sealed trait AttributeValue { self =>
-  def decode[A](implicit schema: Schema[A]): Either[String, A] = Decoder(schema)(self)
+  def decode[A](implicit schema: Schema[A]): Either[String, A] = Codec.decoder(schema)(self)
 
   def ===(that: Operand.Size): ConditionExpression = Equals(ValueOperand(self), that)
   def <>(that: Operand.Size): ConditionExpression  = NotEqual(ValueOperand(self), that)
@@ -48,7 +48,7 @@ object AttributeValue {
 
   def apply[A](a: A)(implicit ev: ToAttributeValue[A]): AttributeValue = ev.toAttributeValue(a)
 
-  def encode[A](a: A)(implicit schema: Schema[A]): AttributeValue = Encoder(schema)(a)
+  def encode[A](a: A)(implicit schema: Schema[A]): AttributeValue = Codec.encoder(schema)(a)
 
   implicit val attributeValueToAttributeValue: ToAttributeValue[AttributeValue] = identity(_)
 }
