+++ dynamodb/src/test/scala/zio/dynamodb/codec/MapCodecSpec.scala
@@ -1,6 +1,6 @@
 package zio.dynamodb.codec
 
-import zio.dynamodb.{ AttributeValue, Decoder, Encoder }
+import zio.dynamodb.{ AttributeValue, Codec }
 import zio.schema.{ DeriveSchema, Schema }
 import zio.test.Assertion.{ equalTo, isRight }
 import zio.test._
@@ -22,12 +22,12 @@ object MapCodecSpec extends DefaultRunnableSpec {
     suite("Map codecs")(
       suite("when encoding")(
         test("encodes map with string key natively") {
-          val actual: AttributeValue = Encoder(HasMapWithStringKey.schema)(HasMapWithStringKey(Map("1" -> 1)))
+          val actual: AttributeValue = Codec.encoder(HasMapWithStringKey.schema)(HasMapWithStringKey(Map("1" -> 1)))
 
           assert(actual.toString)(equalTo("Map(Map(String(map) -> Map(Map(String(1) -> Number(1)))))"))
         },
         test("encodes map with non string key as sequence of Tuple2") {
-          val actual: AttributeValue = Encoder(HasMapWithNonStringKey.schema)(HasMapWithNonStringKey(Map(1 -> 1)))
+          val actual: AttributeValue = Codec.encoder(HasMapWithNonStringKey.schema)(HasMapWithNonStringKey(Map(1 -> 1)))
 
           assert(actual.toString)(
             equalTo("Map(Map(String(map) -> List(Chunk(List(Chunk(Number(1),Number(1)))))))")
@@ -44,7 +44,7 @@ object MapCodecSpec extends DefaultRunnableSpec {
             )
           )
 
-          val actual = Decoder(HasMapWithStringKey.schema)(av)
+          val actual = Codec.decoder(HasMapWithStringKey.schema)(av)
 
           assert(actual)(isRight(equalTo(HasMapWithStringKey(Map("one" -> 1)))))
         },
@@ -59,7 +59,7 @@ object MapCodecSpec extends DefaultRunnableSpec {
             )
           )
 
-          val actual = Decoder(HasMapWithNonStringKey.schema)(av)
+          val actual = Codec.decoder(HasMapWithNonStringKey.schema)(av)
 
           assert(actual)(isRight(equalTo(HasMapWithNonStringKey(Map(1 -> 1)))))
         }
